fix: rename headers & footers components
