+++ src/components/shared/footer/footer.jsx
@@ -1,48 +1,44 @@
 import React from 'react';
 
-import Container from 'components/shared/container';
-import Link from 'components/shared/link';
-import LinkedInIcon from 'icons/linkedin.inline.svg';
-import TwitterIcon from 'icons/twitter.inline.svg';
-import Logo from 'images/logo.inline.svg';
+import leftShape from './images/left-shape.svg';
+import rightShape from './images/right-shape.svg';
 
+const navigation = {
+  main: [
+    { name: 'What is eBPF?', href: 'https://ebpf.io/what-is-ebpf' },
+    { name: 'Blog', href: 'https://ebpf.io/blog' },
+    { name: 'Projects', href: 'https://ebpf.io/projects' },
+    { name: 'Slack', href: 'https://ebpf.io/contribute' },
+  ],
+};
 const Footer = () => (
-  <footer className="py-14 lg:py-8 md:py-6" aria-labelledby="footerHeading">
-    <h2 id="footerHeading" className="sr-only">
-      Footer
-    </h2>
-    <Container className="flex items-center justify-between">
-      <Link to="https://ebpf.io/">
-        <Logo className="lg:w-auto lg:h-8" />
-      </Link>
-
-      <Link
-        className="font-bold transition-colors duration-200 text-primary-2 hover:text-black"
-        to="https://docs.google.com/document/u/0/d/1pmplQdjghadioJVHaGS5X6KrecbYLWVA2VgVz0BzG9E/edit"
-        target="_blank"
-        rel="noopener noreferrer"
-      >
-        Code of Conduct
-      </Link>
-      <div className="flex items-center space-x-6 md:space-x-3">
-        <Link
-          className="text-black transition-colors duration-200 hover:text-gray-1"
-          to="http://twitter.com/ebpfsummit"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <TwitterIcon />
-        </Link>
-        <Link
-          className="text-black transition-colors duration-200 hover:text-gray-1"
-          to="http://linkedin.com/company/ebpf-summit"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <LinkedInIcon />
-        </Link>
-      </div>
-    </Container>
+  <footer className="relative overflow-hidden">
+    <img
+      className="absolute bottom-0 left-0 h-full xl:-left-1/4 md:hidden"
+      src={leftShape}
+      alt=""
+      aria-hidden
+    />
+    <img
+      className="absolute bottom-0 right-0 h-full xl:-right-1/4 md:hidden"
+      src={rightShape}
+      alt=""
+      aria-hidden
+    />
+    <div className="px-4 py-12 mx-auto overflow-hidden max-w-7xl sm:px-6 lg:px-8">
+      <nav className="flex flex-wrap justify-center -mx-5 -my-2" aria-label="Footer">
+        {navigation.main.map((item) => (
+          <div key={item.name} className="px-5 py-2">
+            <a
+              href={item.href}
+              className="font-bold text-black transition-colors duration-200 hover:text-gray-1"
+            >
+              {item.name}
+            </a>
+          </div>
+        ))}
+      </nav>
+    </div>
   </footer>
 );
