--- src/components/shared/main-footer/index.js
@@ -1,1 +1,0 @@
-export { default } from './main-footer';
