--- src/components/shared/main-header/index.js
@@ -1,1 +1,0 @@
-export { default } from './main-header';
