+++ src/components/shared/summit-footer/index.js
@@ -1,0 +1,1 @@
+export { default } from './summit-footer';
