+++ src/components/shared/summit-header/index.js
@@ -1,0 +1,1 @@
+export { default } from './summit-header';
